+++ src/roq/okx/protocol/json/encoder.hpp
@@ -53,7 +53,7 @@ struct Encoder final {
       std::string &buffer,
       CancelAllOrders const &,
       std::string_view const &request_id,
-      std::span<std::pair<std::string_view, std::string_view>> const &symbol_and_external_order_id);
+      std::span<std::pair<int32_t, std::string_view>> const &external_security_id_and_order_id);
 };
 
 }  // namespace json